Refactored test_resources.py into test_modelresources package (#1733)
* Split test_resources.py into test_resources package

* Modified ClasName from Tests to Test Rename some filename

* updated changelog

* merge test updates from release-4

---------
