+++ tests/core/tests/test_resources/test_modelresource/test_resource_factory.py
@@ -1,0 +1,11 @@
+from core.models import Book
+from django.test import TestCase
+
+from import_export import resources
+
+
+class ModelResourceFactoryTest(TestCase):
+    def test_create(self):
+        BookResource = resources.modelresource_factory(Book)
+        self.assertIn("id", BookResource.fields)
+        self.assertEqual(BookResource._meta.model, Book)
